requirements.txt
requests
fastapi
slowapi
pydantic
pytz
tortoise-orm